new extended board class for card handling
